+++ src/index.ts
@@ -1,15 +1,26 @@
-import type { Plugin } from 'vite';
+import type { HtmlTagDescriptor, Plugin } from 'vite';
 import type { MetaConfig } from './interface';
 import { tagsGenerator } from './generator.js';
 
 export type { MetaConfig };
-export function meta(meta: MetaConfig): Plugin {
+
+export function meta(meta: MetaConfig, custom?: (() => HtmlTagDescriptor[]) | HtmlTagDescriptor[] | undefined): Plugin {
 	return {
 		name: 'vite:meta-tags',
 		transformIndexHtml: {
 			enforce: 'post',
 			transform: function () {
-				return Array.from(tagsGenerator(meta));
+				const tags = Array.from(tagsGenerator(meta));
+
+				if (!custom) {
+					return tags;
+				}
+
+				if (typeof custom === 'function') {
+					return [...tags, ...custom()];
+				}
+
+				return [...tags, ...custom];
 			},
 		},
 	};