+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.1.1
+
+- Basic support for custom tags
+- Out-of-the-box support for `google-site-verification` meta tag
+
 ## v0.1.0
 
 - `meta` plugin for injecting meta tags